update chat render
