+++ frontend/pages/_app.js
@@ -1,4 +1,4 @@
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import { Provider, useDispatch } from 'react-redux';
 import { store } from '../store';
 import '@/styles/globals.css';
@@ -11,6 +11,7 @@ import {
 } from '@/store/slices/authSlice';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
+import Script from 'next/script';
 
 function AppContent({ Component, pageProps }) {
   const dispatch = useDispatch();
@@ -40,9 +41,47 @@ function AppContent({ Component, pageProps }) {
   );
 }
 
+function LoadAkvoRag() {
+  const [shouldLoadChat, setShouldLoadChat] = useState(false);
+
+  useEffect(() => {
+    if (typeof window !== 'undefined') {
+      const params = new URLSearchParams(window.location.search);
+      const accessKey = params.get('rag_access');
+      if (accessKey === 'RAG_qnaWr4g') {
+        setShouldLoadChat(true);
+      }
+    }
+  }, []);
+
+  return (
+    <>
+      +{/* Load AkvoRAG JS */}
+      {shouldLoadChat && (
+        <Script
+          src="https://cdn.jsdelivr.net/npm/akvo-rag-js@1.1.5/dist/akvo-rag.js"
+          strategy="afterInteractive"
+          onLoad={() => {
+            if (typeof window !== 'undefined' && window.AkvoRAG) {
+              window.AkvoRAG.initChat({
+                title: 'Kenya Drylands Assistant',
+                kb_id: 36,
+                wsURL: 'ws://akvo-rag.akvotest.org/ws/chat',
+              });
+            } else {
+              console.error('AkvoRAG is not available on window');
+            }
+          }}
+        />
+      )}
+    </>
+  );
+}
+
 function MyApp({ Component, pageProps }) {
   return (
     <Provider store={store}>
+      <LoadAkvoRag />
       <AppContent Component={Component} pageProps={pageProps} />
     </Provider>
   );
